detector/requirements.txt
minio==7.2.0
python-dotenv==1.0.0
ultralytics==8.1.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
torch==2.1.0
torchvision==0.16.0

